+++ Blurred-ios/AccountSettingsViewController.swift
@@ -10,6 +10,9 @@ import UIKit
 
 class AccountSettingsViewController: UIViewController {
 
+    @IBAction func backButton(_ sender: Any) {
+        self.dismiss(animated: true, completion: nil)
+    }
     override func viewDidLoad() {
         super.viewDidLoad()